Various grammar, consistency and unit test fixes
- Changed "Bea Knol" to "Bea Knoller" and "Ad Baantjer" to "Ad Baan"
(See note 1)
- Changed "Sort_by_last_name()" to "Sort_By_Last_Name()" so that
capitalization is now consistent
- Changed "Last_Name_Longer_Then_5()" to "Last_Name_Longer_Than_5()"
(Grammar)
- Changed "LastNameLongerThen(int num)" to "LastNameLongerThan(int num)"
(Grammar)
- Switched "Ad" to "Bea" in the Last_Name_Longer_Than_5 Unit Test so
that it still works after changing the last names of Ad and Bea. (See
change #1)

Note 1:
The last Unit Test (Sort_By_Last_Name_Length()) was also valid when the
student simply returned the unmodified list of names, because the result
of sorting the original list by their last name's lengths was the same
as the order in which they were added to the list.